# src/main/java/org/example/lab1/DTO/OrderEntryDTO.java
package org.example.lab1.DTO;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder(toBuilder = true)
public class OrderEntryDTO {
    private Long orderId;
    private Long productId;
    private int amount;
}